refactor: move bulk of mneme.watch task into `Mneme.Watch.TestRunner`
test: add tests for `Mneme.Watch.ElixirFiles`
